Configurable Configuation Phase (#505)
I have received frequent requests regarding stripping the configuration phase. Despite this, I have never found it sensible to implement or merge this under the condition that users are effectively unable to CHOOSE whether they'd like it stripped (whether it be for compatibility reasons or not). Now that this has been extensively tested and verified to improve the speed at which users swap instances, I can confidently say that this feature was and is ready to merge. If you have any problems, please don't hesitate to reach out.

---------

Co-authored-by: xXkguyXx <12438232+xXkguyXx@users.noreply.github.com>
Co-authored-by: Oleksandr Semenishchev <38401696+semenishchev@users.noreply.github.com>